refactor(webapp): restructure codebase with modular architecture
- **Component Organization**: Split monolithic App.tsx into modular page components
  - Pages: Dashboard, Animations, Settings with dedicated subfolders
  - Components: ButtonPage, CardPage, TablePage with individual variant files
  - Layout: Extracted MainLayout component with sidebar configuration

- **State Management**: Implemented custom hooks for clean separation of concerns
  - usePageNavigation: Handle page routing state
  - useButtonLoading: Manage button loading demonstrations
  - useAnimationReset: Control animation reset functionality
  - useSidebar: Prepare for sidebar state management

- **Data Layer**: Centralized shared data and configuration
  - tableData.tsx: Sample data and column definitions for table component
  - Moved table configuration out of App.tsx for reusability

- **File Structure**: Organized with clear hierarchical patterns
  - src/components/layout/ - Layout components with index exports
  - src/pages/ - Page components with nested subpage structure
  - src/hooks/ - Custom React hooks for state management
  - src/data/ - Shared data and configuration

- **Maintainability Improvements**:
  - Reduced App.tsx from 1200+ to ~60 lines
  - Each component now has single responsibility
  - Consistent index.ts exports for clean imports
  - Type-safe props interfaces for all components

This refactor maintains all existing functionality while significantly improving
code organization, reusability, and developer experience.
