+++ webapp/src/pages/Animations/index.ts
@@ -1,0 +1,1 @@
+export { default as AnimationsPage } from './AnimationsPage';
