+++ webapp/src/pages/Components/subpages/ButtonPage/index.ts
@@ -1,0 +1,5 @@
+export { default as ButtonPage } from './ButtonPage';
+export { default as ButtonVariants } from './ButtonVariants';
+export { default as ButtonSizes } from './ButtonSizes';
+export { default as ButtonIcons } from './ButtonIcons';
+export { default as ButtonExamples } from './ButtonExamples';
