+++ webapp/src/pages/Components/subpages/CardPage/index.ts
@@ -1,0 +1,3 @@
+export { default as CardPage } from './CardPage';
+export { default as CardVariants } from './CardVariants';
+export { default as CardExamples } from './CardExamples';
