+++ webapp/src/pages/Components/subpages/TablePage/TablePage.tsx
@@ -1,0 +1,16 @@
+import TableDemo from './TableDemo';
+import TableConfig from './TableConfig';
+
+const TablePage = () => {
+  return (
+    <div className="showcase-section">
+      <h1>Data Tables</h1>
+      <p>Feature-rich table component with sorting, custom rendering, and responsive design.</p>
+      
+      <TableDemo />
+      <TableConfig />
+    </div>
+  );
+};
+
+export default TablePage;
