+++ webapp/src/pages/Components/subpages/TablePage/index.ts
@@ -1,0 +1,3 @@
+export { default as TablePage } from './TablePage';
+export { default as TableDemo } from './TableDemo';
+export { default as TableConfig } from './TableConfig';
