+++ webapp/src/pages/Components/subpages/index.ts
@@ -1,0 +1,3 @@
+export * from './ButtonPage';
+export * from './CardPage';
+export * from './TablePage';
