+++ webapp/src/pages/Dashboard/index.ts
@@ -1,0 +1,3 @@
+export { default as Dashboard } from './Dashboard';
+export { default as WelcomeSection } from './WelcomeSection';
+export { default as GettingStartedSection } from './GettingStartedSection';
